Add initial implementation and micro app
