+++ microapps/MapViewCalloutApp/app/src/main/java/com/arcgismaps/toolkit/mapviewcalloutapp/screens/MainScreen.kt
@@ -1,0 +1,64 @@
+/*
+ *
+ *  Copyright 2024 Esri
+ *
+ *  Licensed under the Apache License, Version 2.0 (the "License");
+ *  you may not use this file except in compliance with the License.
+ *  You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ *  Unless required by applicable law or agreed to in writing, software
+ *  distributed under the License is distributed on an "AS IS" BASIS,
+ *  WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ *  See the License for the specific language governing permissions and
+ *  limitations under the License.
+ *
+ */
+
+package com.arcgismaps.toolkit.mapviewcalloutapp.screens
+
+import androidx.compose.foundation.layout.fillMaxSize
+import androidx.compose.runtime.Composable
+import androidx.compose.runtime.getValue
+import androidx.compose.runtime.mutableStateOf
+import androidx.compose.runtime.remember
+import androidx.compose.ui.Modifier
+import androidx.compose.material3.Text
+import androidx.compose.ui.graphics.Color
+import com.arcgismaps.geometry.Point
+import com.arcgismaps.geometry.SpatialReference
+import com.arcgismaps.mapping.ArcGISMap
+import com.arcgismaps.mapping.BasemapStyle
+import com.arcgismaps.mapping.Viewpoint
+import com.arcgismaps.toolkit.geoviewcompose.Callout
+import com.arcgismaps.toolkit.geoviewcompose.MapView
+import com.arcgismaps.toolkit.geoviewcompose.MapViewScope
+
+@Composable
+fun MainScreen() {
+    val arcGISMap by remember {
+        mutableStateOf(
+            ArcGISMap(BasemapStyle.ArcGISTopographic).apply {
+                initialViewpoint = Viewpoint(
+                    latitude = 39.8,
+                    longitude = -98.6,
+                    scale = 10e7
+                )
+            }
+        )
+    }
+
+    val disneyLand = remember { Point(-117.9190, 33.8121, SpatialReference.wgs84()) }
+    val callout: (@Composable MapViewScope.() -> Unit)? by remember {
+        mutableStateOf(
+            { Callout(location = disneyLand) { Text("Hello, World!", color = Color.Green) } }
+        )
+    }
+
+    MapView(
+        modifier = Modifier.fillMaxSize(),
+        arcGISMap = arcGISMap,
+        content = callout
+    )
+}
