--- microapps/TemplateApp/app/src/main/java/com/arcgismaps/toolkit/templateapp/MainActivity.kt
@@ -1,55 +1,0 @@
-/*
- *
- *  Copyright 2024 Esri
- *
- *  Licensed under the Apache License, Version 2.0 (the "License");
- *  you may not use this file except in compliance with the License.
- *  You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- *  Unless required by applicable law or agreed to in writing, software
- *  distributed under the License is distributed on an "AS IS" BASIS,
- *  WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- *  See the License for the specific language governing permissions and
- *  limitations under the License.
- *
- */
-
-package com.arcgismaps.toolkit.templateapp
-
-import android.os.Bundle
-import androidx.activity.ComponentActivity
-import androidx.activity.compose.setContent
-import androidx.compose.runtime.Composable
-import androidx.compose.ui.tooling.preview.Preview
-import com.arcgismaps.ApiKey
-import com.arcgismaps.ArcGISEnvironment
-import com.arcgismaps.toolkit.templateapp.screens.MainScreen
-import com.arcgismaps.toolkit.templateapp.ui.theme.TemplateAppTheme
-
-class MainActivity : ComponentActivity() {
-    override fun onCreate(savedInstanceState: Bundle?) {
-        super.onCreate(savedInstanceState)
-        ArcGISEnvironment.apiKey =
-            ApiKey.create(BuildConfig.API_KEY)
-        setContent {
-            TemplateAppTheme {
-                TemplateApp()
-            }
-        }
-    }
-}
-
-@Composable
-fun TemplateApp() {
-    MainScreen()
-}
-
-@Preview(showBackground = true)
-@Composable
-fun AppPreview() {
-    TemplateAppTheme {
-        TemplateApp()
-    }
-}
