+++ settings.gradle.kts
@@ -88,6 +88,8 @@ include(":floor-filter-app")
 project(":floor-filter-app").projectDir = File(rootDir, "microapps/FloorFilterApp/app")
 include(":geoview-compose")
 project(":geoview-compose").projectDir = File(rootDir, "toolkit/geoview-compose")
+include(":map-view-callout-app")
+project(":map-view-callout-app").projectDir = File(rootDir, "microapps/mapviewcalloutapp/app")
 include(":map-view-location-display-app")
 project(":map-view-location-display-app").projectDir = File(rootDir, "microapps/mapviewlocationdisplayapp/app")
 include(":map-view-insets-app")
