+++ toolkit/geoview-compose/src/main/java/com/arcgismaps/toolkit/geoviewcompose/Callout.kt
@@ -1,0 +1,83 @@
+/*
+ *  Copyright 2024 Esri
+ *
+ *  Licensed under the Apache License, Version 2.0 (the "License");
+ *  you may not use this file except in compliance with the License.
+ *  You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ *  Unless required by applicable law or agreed to in writing, software
+ *  distributed under the License is distributed on an "AS IS" BASIS,
+ *  WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ *  See the License for the specific language governing permissions and
+ *  limitations under the License.
+ *
+ */
+
+package com.arcgismaps.toolkit.geoviewcompose
+
+import androidx.compose.foundation.BorderStroke
+import androidx.compose.foundation.background
+import androidx.compose.foundation.border
+import androidx.compose.foundation.layout.Box
+import androidx.compose.foundation.layout.offset
+import androidx.compose.foundation.layout.wrapContentSize
+import androidx.compose.material3.MaterialTheme
+import androidx.compose.runtime.Composable
+import androidx.compose.runtime.collectAsState
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.graphics.Color
+import androidx.compose.ui.platform.LocalDensity
+import androidx.compose.ui.unit.dp
+import com.arcgismaps.geometry.Point
+import com.arcgismaps.mapping.view.DrawStatus
+import com.arcgismaps.mapping.view.MapView
+import com.arcgismaps.mapping.view.ScreenCoordinate
+
+public class MapViewScope(internal val mapView: MapView)
+
+/**
+ * Creates a Callout at the specified geographical location on the MapView. The Callout is a composable
+ * that can be used to display additional information about a location on the map. The additional information is
+ * passed as a content composable that contains text and/or other content. It has a leader that points to
+ * the location that Callout refers to. The body of the Callout is a rectangular area with curved corners
+ * that contains the content lambda provided by the application. A thin border line is drawn around the entire Callout.
+ *
+ * @param location the geographical location at which to display the Callout
+ * @param modifier Modifier to be applied to the composable Callout
+ * @param content the content of the Callout
+ * @since 200.5.0
+ */
+@Composable
+public fun MapViewScope.Callout(location: Point,
+                                modifier: Modifier = Modifier,
+                                content: @Composable () -> Unit) {
+
+    if (mapView.drawStatus.collectAsState().value == DrawStatus.InProgress) return
+
+    val calloutScreenCoordinate: ScreenCoordinate = mapView.locationToScreen(location)
+    Box(
+        modifier = Modifier
+            .offset(
+                x = with(LocalDensity.current) {
+                    calloutScreenCoordinate.x
+                        .toFloat()
+                        .toDp()
+                },
+                y = with(LocalDensity.current) {
+                    calloutScreenCoordinate.y
+                        .toFloat()
+                        .toDp()
+                })
+            .wrapContentSize()
+            .background(Color.White)
+            .border(
+                border = BorderStroke(2.dp, Color.LightGray),
+                shape = MaterialTheme.shapes.medium
+            )
+    )
+    {
+        content.invoke()
+    }
+}
